backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
PyMuPDF==1.24.1
python-docx==1.1.2
pandas==2.2.2
reportlab==4.2.0
langdetect==1.0.9
spacy==3.7.4
tenacity==8.3.0
aiofiles==23.2.1